remove warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@20834 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/kernel/discrete/signalprocessor/voronoi/representation/voronoicell/BestVoronoiCells.java
@@ -51,7 +51,7 @@ public class BestVoronoiCells {
         double sum = 0;
         if ( numberstored==0 ) return -1.0;
         for ( int x=0 ; x<numberstored ; x++ ) sum += bestareas[x];
-        return sum/((double)numberstored);
+        return sum/(numberstored);
     }
     
     public int getAverageX() {
