+++ common/src/org/apache/commons/math/util/MathUtils.java
@@ -2215,7 +2215,7 @@ public final class MathUtils {
     double s3=0.0;
     double x1max = 0.0;
     double x3max = 0.0;
-    double floatn = (double)v.length;
+    double floatn = v.length;
     double agiant = rgiant/floatn;
     for (int i=0;i<v.length;i++) {
         double xabs = Math.abs(v[i]);
